wsb-jpa1/src/main/test/com/capgemini/wsb/persistance/dao/VisitServiceTest.java
package com.capgemini.wsb.service.impl;

import com.capgemini.wsb.persistence.entity.VisitEntity;
import com.capgemini.wsb.service.VisitService;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import static org.assertj.core.api.Assertions.assertThat;

import java.util.List;

@SpringBootTest
public class VisitServiceTest {

    @Autowired
    private VisitService visitService;

    @Test
    public void testFindVisitsByPatientId() {
        List<VisitEntity> visits = visitService.findVisitsByPatientId(1L);
        assertThat(visits).isNotEmpty();
    }
}
